Whats new: (From big break) 1.RealWorkHour/Index now shows employees that worked but dont exist in the timeshift 2.Projection RealWorkHoursSpecificDates fixed 3.Projection RealWorkHoursAnalytically fixed 4.Projection RealWorkHoursAnalyticallySum fixed 5.Make excel WorkHours go brrrrrr 6.Projection RealWorkHoursAnalyticallySum fixed 7. Προεραιρικα -> Προαιρετικα 8.Projection EmployeeConsecutiveDayOff fixed 9.Security repository small changes(Same as base) 10.New JS service AppendListService() 11.Add list of Αρχιφυλακες in workplace edit/details 12.Add Logs with 3 tables and a service 13. realworkhours createdOn now works on create

+++ DataAccess/ViewModels/Entities/Contracts/ContractCreate.cs
@@ -17,7 +17,7 @@ namespace DataAccess.ViewModels
         public int HoursPerWeek { get; set; }
 
         [Display(Name = "Ώρες ανα ημέρα")]
-        public int HoursPerDay { get; set; }
+        public decimal HoursPerDay { get; set; }
 
         [Display(Name = "Εργάσιμες μέρες εβδομάδας")]
         public int WorkingDaysPerWeek { get; set; }
